requirements.txt
flask
pyotp
bcrypt
python-dotenv
qrcode[pil]
pillow
